# feat: add data type conversion for generic encoding/decoding unions