## Goal of this PR

<!-- A brief description of the change being made with this pull request. -->
Extends #515 to allow type conversions for Avro Unions when encoding/decoding Go `any`.

## How did I test it?

Added more unit tests (and split the unit test file into separate files for encode and decode).

Benchmark comparison (10 runs each) looks good; any differences are just noise:
```
                          │   old.txt    │               new.txt                │
                          │    sec/op    │    sec/op     vs base                │
SuperheroDecode-12          595.6n ±  7%   626.1n ±  8%        ~ (p=0.063 n=10)
SuperheroEncode-12          612.3n ±  7%   508.7n ± 27%  -16.91% (p=0.043 n=10)
PartialSuperheroDecode-12   289.9n ± 16%   262.2n ± 13%   -9.54% (p=0.011 n=10)
SuperheroGenericDecode-12   8.184µ ± 29%   9.004µ ± 10%        ~ (p=0.143 n=10)
SuperheroGenericEncode-12   1.054µ ±  5%   1.030µ ± 12%        ~ (p=0.565 n=10)
SuperheroWriteFlush-12      443.2n ± 17%   416.7n ± 22%        ~ (p=0.280 n=10)
geomean                     859.9n         828.2n         -3.69%
```